cleaning up file structure
