+++ examples/access/create.yaml
@@ -7,19 +7,23 @@ Tests:
   - Name: Create Project, App, and User
     Projects:
       - Name: e2e-test-access-project%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Test: C
     Applications:
       - Name: e2e-test-access-app%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Rules: 
           - Type: "project.name.in"
             Value: "e2e-test-access-project%E2E_RUN_SUFFIX%"
         Test: C
     Groups:
       - Name: e2e-test-access-group%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Roles: [ ast-scanner ]
         Test: C
     Users:
       - Name: e2e-test-access-user%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Email: e2e_test_access_user1@cx.local
         Roles: [ ast-scanner ]
         Test: C
@@ -27,6 +31,7 @@ Tests:
     AccessAssignments:
       - EntityType: user
         EntityName: e2e-test-access-user%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         ResourceType: project
         ResourceName: e2e-test-access-project%E2E_RUN_SUFFIX%
         Roles: [ ast-scanner ]
@@ -35,6 +40,7 @@ Tests:
     AccessAssignments:
       - EntityType: user
         EntityName: e2e-test-access-user%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         ResourceType: project
         ResourceName: e2e-test-access-project%E2E_RUN_SUFFIX%
         Roles: [ ast-scanner ]
