+++ examples/access/delete.yaml
@@ -7,13 +7,17 @@ Tests:
   - Name: App, Project, and Group Delete
     Projects:
       - Name: e2e-test-access-project%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Test: RD
     Applications:
       - Name: e2e-test-access-app%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Test: RD
     Users:
       - Name: e2e-test-access-user%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Test: RD
     Groups:
       - Name: e2e-test-access-group%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "ACCESS_MANAGEMENT_ENABLED" ]
         Test: RD
