+++ examples/query/app.yaml
@@ -9,6 +9,7 @@ Tests:
       - Name: Client_DOM_XSS
         Group: JavaScript_High_Risk
         Language: JavaScript
+        FeatureFlags: [ "AUDIT_APPLICATION_LEVEL_ENABLED" ]
         Scope: 
           Tenant: false
           Application: e2e-test-query-app%E2E_RUN_SUFFIX%
@@ -22,12 +23,14 @@ Tests:
         ZipFile: ../files/xss-burger.zip
         Branch: zip
         Preset: e2e-test-query-preset%E2E_RUN_SUFFIX%
+        FeatureFlags: [ "AUDIT_APPLICATION_LEVEL_ENABLED" ]
         Engine: sast
         WaitForEnd: true
         Test: C
     Results:
       - Project: e2e-test-query-project%E2E_RUN_SUFFIX%
         FindingNumber: 1 # this finding should be gone again
+        FeatureFlags: [ "AUDIT_APPLICATION_LEVEL_ENABLED" ]
         Filter:
           Query: Client_DOM_XSS
           Severity: HIGH
@@ -38,6 +41,7 @@ Tests:
       - Name: Client_DOM_XSS
         Group: JavaScript_High_Risk
         Language: JavaScript
+        FeatureFlags: [ "AUDIT_APPLICATION_LEVEL_ENABLED" ]
         Scope: 
           Tenant: false
           Application: e2e-test-query-app%E2E_RUN_SUFFIX%
