+++ pkg/types/groups.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"fmt"
@@ -8,12 +8,21 @@ import (
 )
 
 func (t *GroupCRUD) Validate(CRUD string) error {
+	if (CRUD == OP_UPDATE || CRUD == OP_DELETE) && t.Group == nil {
+		return fmt.Errorf("must read before updating or deleting")
+	}
+
 	if t.Name == "" {
 		return fmt.Errorf("group name is missing")
 	}
 
 	return nil
 }
+
+func (t *GroupCRUD) IsSupported(CRUD string) bool {
+	return true
+}
+
 func (t *GroupCRUD) GetModule() string {
 	return MOD_GROUP
 }
