+++ pkg/types/presets.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"fmt"
@@ -8,12 +8,21 @@ import (
 )
 
 func (t *PresetCRUD) Validate(CRUD string) error {
+	if (CRUD == OP_UPDATE || CRUD == OP_DELETE) && t.Preset == nil {
+		return fmt.Errorf("must read before updating or deleting")
+	}
+
 	if t.Name == "" {
 		return fmt.Errorf("preset name is missing")
 	}
 
 	return nil
 }
+
+func (t *PresetCRUD) IsSupported(CRUD string) bool {
+	return true
+}
+
 func (t *PresetCRUD) GetModule() string {
 	return MOD_PRESET
 }
