+++ pkg/types/projects.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"fmt"
@@ -8,12 +8,21 @@ import (
 )
 
 func (t *ProjectCRUD) Validate(CRUD string) error {
+	if (CRUD == OP_UPDATE || CRUD == OP_DELETE) && t.Project == nil {
+		return fmt.Errorf("must read before updating or deleting")
+	}
+
 	if t.Name == "" {
 		return fmt.Errorf("project name is missing")
 	}
 
 	return nil
 }
+
+func (t *ProjectCRUD) IsSupported(CRUD string) bool {
+	return true
+}
+
 func (t *ProjectCRUD) GetModule() string {
 	return MOD_PROJECT
 }
