+++ pkg/types/roles.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"fmt"
@@ -8,12 +8,21 @@ import (
 )
 
 func (t *RoleCRUD) Validate(CRUD string) error {
+	if (CRUD == OP_UPDATE || CRUD == OP_DELETE) && t.Role == nil {
+		return fmt.Errorf("must read before updating or deleting")
+	}
+
 	if t.Name == "" {
 		return fmt.Errorf("role name is missing")
 	}
 
 	return nil
 }
+
+func (t *RoleCRUD) IsSupported(CRUD string) bool {
+	return true
+}
+
 func (t *RoleCRUD) GetModule() string {
 	return MOD_ROLE
 }
